refactor(gitgateway): move testutil since used in other package
