+++ gitgateway/v2/git_test.go
@@ -2,41 +2,14 @@ package gitgateway
 
 import (
 	"bou.ke/monkey"
-	"os"
 	"testing"
 	"time"
 
-	"github.com/go-git/go-billy/v5/osfs"
-	"github.com/go-git/go-billy/v5/util"
-	"github.com/go-git/go-git/v5"
 	"github.com/stretchr/testify/assert"
 )
 
-func temporalDir() (path string, clean func()) {
-	fs := osfs.New(os.TempDir())
-	path, err := util.TempDir(fs, "", "")
-	if err != nil {
-		panic(err)
-	}
-	return fs.Join(fs.Root(), path), func() {
-		err := util.RemoveAll(fs, path)
-		if err != nil {
-			panic(err)
-		}
-	}
-}
-
-func initRepository() (path string, clean func()) {
-	dir, clean := temporalDir()
-	_, initErr := git.PlainInit(dir, false)
-	if initErr != nil {
-		panic(initErr)
-	}
-	return dir, clean
-}
-
 func TestGitGateway(t *testing.T) {
-	path, clean := initRepository()
+	path, clean := InitRepository()
 	defer clean()
 
 	git, err := NewGitGatewayLocal(path)
@@ -89,7 +62,7 @@ func TestGitGateway(t *testing.T) {
 func TestCommitDuration(t *testing.T) {
 	defer monkey.UnpatchAll()
 
-	path, clean := initRepository()
+	path, clean := InitRepository()
 	defer clean()
 
 	git, err := NewGitGatewayLocal(path)
